refactor: extract contrast computations into helpers.ts
